+++ src/DataCliente.cpp
@@ -1,4 +1,4 @@
-#include "../../include/DTs/DataCliente.h"
+#include "../include/DataCliente.h"
 
 DataCliente::DataCliente(string _nickname, DTFecha* _fecha, DTDireccion* _direccion, string _ciudad):
 DataUsuario(_nickname, _fecha), direccion(_direccion), ciudad(_ciudad){};
