+++ src/DataComentario.cpp
@@ -1,0 +1,19 @@
+#include "../include/DataComentario.h"
+
+DataComentario::DataComentario(string _comentario, DTFecha* _fecha, int id){
+    comentario = _comentario;
+    fecha = _fecha;
+    idComentario = id;
+}
+
+DataComentario::~DataComentario(){};
+
+string DataComentario::getComentario(){
+    return comentario;
+}
+DTFecha* DataComentario::getFecha(){
+    return fecha;
+}
+int DataComentario::getIdComentario(){
+    return idComentario;
+}
