+++ src/DataVendedor.cpp
@@ -1,4 +1,4 @@
-#include "../include/DTs/DataVendedor.h"
+#include "../include/DataVendedor.h"
 
 DataVendedor::DataVendedor(string _nickname, DTFecha* _fecha, int _RUT):
 DataUsuario(_nickname, _fecha), RUT(_RUT){};
