+++ src/SuscribirseANotificaciones.cpp
@@ -1,6 +1,6 @@
 #include <string>
 #include <set>
-#include "../include/DTs/DataVendedor.h"
+#include "../include/DataVendedor.h"
 using namespace std;
 
 void SuscribirseANotficaciones(string nicknameCliente){
